timer: fixing issues caused by slot processing being non-atomic
One of the refactors I did in the prior intrusive timer commit made it so that
we would re-register and awaken timers based on their expiration time instead
of blindly dropping them down to the next level. This is all well and good,
except that the slot they need to be registered in depends on the current time,
and when we find a timer that needs expiring halfway through processing a slot,
if we were to bail out from slot processing at that point we'd leave things in
an inconsistent state.

To fix this, we add a temporary queue to hold timers that need expiration that
we find during slot processing; once we're done processing slots, we go back
and start firing off timers from the pending queue.