timer: fix memory model issues
We need stronger synchronization than was previously present to ensure that
future writes from the thread that owns the timer don't appear before the timer
is fired. Relaxed ordering is tricky, eh?

To resolve this, we now acquire the driver lock on every timer drop. In
addition, to simplify reasoning in poll_timeout, we perform waker registration
unconditionally before the state load.