Implement a test scheduler that propagates panics
This patch implements a new scheduler that does not capture unwinds. It
instead lets unwinds bubble up.

This is useful in test cases where we might be interested in detached
tasks' panics. Not all applications are interested in letting tasks
fail, so this scheduler makes the implementation of testing such
applications or libraries much easier.