util: handle removal of expiring DelayQueue entry
The DelayQueue remove and reset methods can panic when attempting to
remove certain keys whose deadlines have been reached but which have not
yet been retrieved using poll_expired.

The panic occurs in the remove method of the timer wheel implementation
ported from the tokio core, which asserts on an attempt to convert from
a millisecond offset to an internal wheel level if given a value where
(a) both the level and slot index would be 0 and (b) the value is equal
to the wheel's current time offset.

That assert would be valid on the assumption that the wheel should not
contain an entry which expires at the wheel's current time, but there
may be multiple wheel entries with the same deadline and only one entry
is removed from the wheel per poll call, with the wheel advancing its
internal time to those entries' deadline as it shifts entries between
levels before popping the first of the entries.

Since the general case where there are multiple entries with the same
deadline is already well handled, it's only necessary to avoid panicking
in Wheel::remove on the special case where the slot index is 0, which we
do by masking the slot bits before computing the level.