DerefPinMut
Introduce `DerefPinMut` trait. It is similar to `DerefMut`, but
dereferences `Pin<&mut Self>` to `Pin<&mut Self::Target>`.

There are two reasons to do that:

* removal of copy-paste and macros in tokio implementation
  (this commit diffstat is 62 insertions(+), 126 deletions(-),
  excluding a new test)
* ability to reuse this delegation implementation in user crates

The latter was my motivation for this commit. I have a couple
Box-like wrappers to traits, and implementing this delegation is
tedious and error-prone.

It tokio implement `AsyncWrite`, `AsyncRead` and friends for
`DerefPinMut`, then I could only implement `DerefPinMut` in my crate
and get all `Async` trait implementations for free.