tokio-stream: WatchStream: yield the current value
This udptes WatchStream to always yield the current value before waiting
for updates. This makes it possible to use it as a drop-in replacement
for watch::Receiver when updating a Tokio 0.2 application.

There is one caveat here, which is that we can't tell if the sender was
used before. This means we might yield the current value twice if it was
(because reading what we think is the initial value doesn't increment
the watcher).

I don't think there really is a way to fix this without changing the API
of Receiver, so for now I went with documenting it. Let me know what you
think.

More context: https://github.com/tokio-rs/tokio/discussions/3575