macros: preserve none-delimiters in function body
When user-generated token streams are returned from proc macros, all
none-delimited groups are flattened. This can cause compilation errors
as shown in the issue below. By instead passing the function body token
stream straight through from the input of the proc macro to its output,
it does not get deconstructed and reconstructed, and so the
none-delimited groups are not lost.

Fixes: #3579