test: Mark the Mock IO object as Send
Previously, the inner 'rx' of the Mock IO object used an unbound
receiver channel for mock 'Actions'. This allowed the Send auto
trait implentation to match this object. Mock was recently
convernted to use dynamic dispatch, which now needs the marker
Send internally inorder for it to also be Send.