Introduce `biased` argument for `select!`
Similar to `futures::select_biased!`, it is sometimes beneficial to
provide developers with strict control over the polling order of
futures. Most of the time, the fairness guarantees that `select!`
provides through RNG can be thought out by a careful developer to ensure
that no one future will monopolize all of the CPU time. In exchange for
this, the developer gets a performance boost as they do not need to
spend time generating endless random numbers.

Fixes: #2181