Document panic in oneshot::Receiver::try_recv
When initially consulting the docs I assumed the channel would report as
closed after the value was sent and received partially due to not
noticing any mention of panics. I had a look through the module and
Receiver docs for oneshot and didn't find mention of the panic so
updated the receiver docs accordingly.