time: allow users to specify `Interval` behavior when delayed
Occasionally, `tokio::time::Interval` may be delayed from ticking by a
blocked executor. By default, `Interval` continues to schedule ticks
normally, which causes a burst of ticks until it is "caught up" in time
where it should be. This behavior is not always desired, however, so
this commit adds the `MissedTickBehavior` enum to the `time` module to
allow users to specify what behavior they'd like `Interval` to exhibit
in such situations.

Closes #3574