refactor: Add a ReadBuf::with_buf
Builds on `ReadBuf::from_buf` to also factor out the duplicated code
when updating the `BufMut` after the read.

The closure could return just `R` instead of `Poll<io::Result<R>>` but
it would then update the `BufMut` regardless of `Ok/Err/Pending` which
does not match the current code.