macros: fix wrong error messages
The macros `tokio::main` and `tokio::test` return invalid error
messages if the function being wrapped has no return value. This was
because the semicolon from the last statement was missing.

Also, since version 0.1.54 (or earlier),
`clippy::semicolon_if_nothing_returned` triggered false-negative when
using the macros `tokio::test` and `tokio::main` on functions without
a result type.

We can copy the semicolon and the `return` keyword from the last
statement from the body of the wrapped function.

Fixes: rust-lang/rust-clippy#7438