Fix the PollAio tests' use of poll()
The original version of these tests used a
futures_test::task::noop_context so it could manually poll the futures.
But it still constructed a Runtime, because PollAio requires a Runtime
for its construction.  This caused the Runtime to poll the future in a
tight loop after event notification.

Remove noop_context, implement Future, and use a normal #[tokio::test]
instead.