make timeout robust against budget-depleting tasks
Up until now, if the future that we were applying a timeout to
consistently depleted the coop budget, the timeout never got a chance to
be evaluated. In the next call to `poll`, the underlying future would be
polled and it would once again deplete the budget. In those circumstances,
timeouts would not be respected. This can be surprising to people, and
in fact it was in #4291 .

The solution is to make a budget exception with `timeout` if it was the
underlying future that depleted the budget.

Refs: #4291 , #4300