wip: try initial solution
The `UnownedTask` type will have an extra field called `is_mandatory`
which will determine whether the tasks get executed at shutdown. Maybe
the `RawTask` is a better place to put this field, but I don't know my
way around these types yet.

I have verified that this code change fixes the problem by running the
code in the original issue a hundred thousand times. Without the fix,
that code consistently misses a write in the first few hundred
executions. With the change, I've never seen it miss the write.

I think we might be able to use `loom` to test this. Will try to do so.