fix build attributes for spawn_mandatory_blocking
When running `cargo test` on `<repo>/tokio`, you would expect that
`rustc` only gets executed against `tokio` using the `--test` flag,
which would make the `cfg_attr(test, ...)` conditional attribute apply.
That's not true though. When running `cargo test` on `tokio`, the tokio
crate gets built *twice*, once without `--test` and once with (this can
be verified using strace). This is because, f.i, tokio specifies a
dev-dependency on `tokio-test, which specifies back a dependency on
`tokio`.

So when running a regular test build, tokio will first get built without
`--test`. We will not get dead code errors there because for that build,
`tokio::fs::File` uses the new `spawn_mandatory_blocking`. For the next
build, the one with `--test`, we will not get dead code errors because,
even though `tokio::fs::File` uses a mock `spawn_mandatory_blocking`, the
`cfg_attr(test` is working as expected.

Things are different for loom builds. We will first get a build of tokio
without `--test` but with the `--cfg loom` flag. The fact that
`tokio::fs::File` uses `spawn_mandatory_blocking` won't save us here,
because `tokio:fs` doesn't get compiled in loom builds

The solution I can think of is extending the `cfg_attr(test` to
`cfg_attr(any(test, loom)`, which is a bit unfortunate because
`spawn_mandatory_blocking` *is* used in the loom tests, only not in the
regular loom build of the crate, which is triggered by the cyclical
dependency between tokio and tokio-test.