io: make `EPOLLERR` awake `AsyncFd::readable`
Currently, if a user is awaiting on a `AsyncFd` becoming readable and we
get a polling error (which may happen in some cases, see the referenced
issue), the awaiter will never be woken. Ideally, they would be notified
of this error in some way.

The solution here involves communicating to the `Readiness` object that
there was an error, and when that happens, the `Readiness` future will
return a synthetic error that boils down to "there was a polling error".
If the user wants to find more information about the underlying error,
they would have to find some other way (not clear how yet? as pointed
out in the issue, for sockets one could use `SO_ERROR`).