spawn_mandatory_blocking: fix race with shutdown
On a previous iteration of this PR, there could be a race between
calling `spawn_mandatory_blocking` and dropping the runtime. This could
result in f.i, a call to `File::write` returning successfully but the
actual write not getting scheduled.

Now `spawn_mandatory_blocking` will keep track of whether the task was
actually scheduled. If not, the `File::write` method will return an
error, letting users know that the write did not, and will not happen.

I have also added a loom test that checks that the return value of
`spawn_mandatory_blocking` can be trusted, even when shutting down from
a different thread.