deprecate reexports from std
Reexports for the sake of it are not helpful (confusing actually)
from std. For new developers, it's not obvious that
tokio::time::Duration is the same as std::time::Duration.

Looking a bit into the history, they appear motivated in an earlier
less stable version of tokio when the types may have been changing
or weren't in std yet.