task: fix unstable API documentation notes
PR #4499 made the `JoinSet` API unstable, but did not add a
documentation note explaining unstable features. This branch adds a note
on unstable APIs to the `JoinSet` type-level documentation, similar to
the notes for `task::Builder` and the runtime metrics APIs.

Also, I noticed that there was a broken link to the top-level
documentation on unstable APIs in the docs for `task::Builder`, so I
fixed that as well.