reduce changes to `IdleNotifiedSet` a lot; simplify implementation
This does, unfortunately, mean that keys cannot be returned with
cancelled tasks, as they are only stored in the map and may have been
evicted when replacing a task...however, we could change this
behavior by not returning cancelled join errors for tasks that were
aborted by replacing in the map. This would mean removing a task from
the `IdleNotifiedSet` as well when it is replaced by `insert`.

On the other hand, this does make the API closer to `JoinSet`'s, even if
it's somewhat less nice.