Fix aliasing problems in CancellationToken{State}
The biggest problem was that code was reborrowing a raw pointer into a
reference, then freeing the reference inside of a method call. This
produces a dangling reference, because reference parameters must still
be valid when the function returns. This is addressed by not converting
to a reference. Unfortunately this means that several methods have to
become associated functions, because *const Self cannot be a receiver
for a method call. So there's an ergonomic hit, but at least the change
is mostly inline.

This also hoists some unsafety out to the function boundary. This seems
appropriate for a function that may free the pointer which it is passed.

This also addresses a different aliasing problem which is only
technically an issue in the presence of raw pointer tagging, where there
was a conflict between use of Box and &mut in
CancellationToken::child_token. Both Box and &mut claim unique ownership
over the pointee upon use, so interleaving them (and generally creating
a NonNull from a &mut) runs afoul of Stacked Borrows with raw pointer
tagging.